requirements.txt
altgraph==0.17
async-lru==1.0.2
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
easyocr==1.3.1
future==0.18.2
idna==2.10
imageio==2.9.0
importlib-metadata==4.0.1
kiwisolver==1.3.1
loguru==0.5.3
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.3
opencv-python==4.5.1.48
pefile==2019.4.18
Pillow==8.1.2
pyaes==1.6.1
pycodestyle==2.7.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.1.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3.0.2
PyQtWebEngine==5.15.2
PyQtWebEngine-Qt5==5.15.2
Pyrogram==1.2.9
PySocks==1.7.1
pytesseract==0.3.7
python-bidi==0.4.2
python-dateutil==2.8.1
python-dotenv==0.16.0
PyWavelets==1.1.1
pywin32==300
pywin32-ctypes==0.2.0
pyWinhook==1.6.2
PyYAML==5.4.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
QtAwesome==1.0.2
QtPy==1.9.0
requests==2.25.1
scikit-image==0.18.1
scipy==1.6.3
six==1.15.0
SpeechRecognition==3.8.1
TgCrypto==1.2.2
tifffile==2021.4.8
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
urllib3==1.26.4
win32-setctime==1.0.3
wxPython==4.1.1
zipp==3.4.1

